refactor: split runners.py into sub-modules to fix 500-line cap
runners.py (1004 lines) was breaking the ci-health-check 500-line gate,
blocking all PR CI pipelines. Split into focused sub-modules:

- runner_helpers.py  (155 lines): shared utility fns
- runner_groups.py   (186 lines): label-based group operations
- runner_diagnostics.py (376 lines): diagnostics, fleet capacity,
  autoscaling, troubleshooting
- runners.py         (336 lines): core lifecycle routes only

Also fix pre-existing ruff I001/UP017 in routers/system.py.

Resolves ci-health-check 'Verify no source file exceeds 500 lines'.
